refactor(ui-components): broke component demo into multiple pages
